+++ client/src/routes/groups/[group_id]/tabs/ExpensesTab.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 	import Button from '$lib/components/ui/Button.svelte';
-	import { formatExpenseDate } from '../group.svelte';
+	import { formatExpenseDate } from '$lib/utils/format_utils';
 	import type { GroupState } from '../group.svelte';
 
 	let { groupState, onCreateExpense } = $props<{
